Extract janitor logic and test nothing to clean scenario
